% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random_Numbers.R
\name{rtelliptical}
\alias{rtelliptical}
\title{Sampling Random Numbers from Truncated Multivariate Elliptical Distributions}
\usage{
rtelliptical(n = 10000, mu = rep(0, length(lower)),
  Sigma = diag(length(lower)), lower, upper = rep(Inf, length(lower)),
  dist = "Normal", nu = NULL, expr = NULL, gFun = NULL,
  ginvFun = NULL, burn.in = 0, thinning = 1)
}
\arguments{
\item{n}{number of observations to generate. Must be an integer >= 1.}

\item{mu}{numeric vector of length \eqn{p} representing the location parameter.}

\item{Sigma}{numeric positive definite matrix with dimension \eqn{p}x\eqn{p} representing the
scale parameter.}

\item{lower}{vector of lower truncation points of length \eqn{p}.}

\item{upper}{vector of upper truncation points of length \eqn{p}.}

\item{dist}{represents the truncated distribution to be used. The values are \code{'Normal'},
\code{'t'}, \code{'PE'}, \code{'PVII'}, \code{'Slash'}, and \code{'CN'} for the truncated Normal, Student-t,
Power Exponential, Pearson VII, Slash, and Contaminated Normal distribution, respectively.}

\item{nu}{additional parameter or vector of parameters depending on the
density generating function. See Details.}

\item{expr}{a character with the density generating function. See Details.}

\item{gFun}{an R function with the density generating function. See Details.}

\item{ginvFun}{an R function with the inverse of the density generating function defined in
\code{gFun}. See Details.}

\item{burn.in}{number of samples to be discarded as a burn-in phase.}

\item{thinning}{factor for reducing the autocorrelation of random points.}
}
\value{
It returns a matrix of dimensions \eqn{n}x\eqn{p} with the random points sampled.
}
\description{
The sampling procedure is based on a Slice Sampling algorithm combined with
Gibbs sampling steps. The distribution is characterized by a location parameter
\code{mu}, a scale matrix \code{Sigma}, and lower and upper truncation bounds
specified by \code{lower} and \code{upper}, respectively.
}
\details{
The argument \code{dist} specifies the truncated distribution to be used and
accepts the values \code{"Normal"}, \code{"t"}, \code{"PE"}, \code{"PVII"},
\code{"Slash"}, and \code{"CN"}, corresponding to the truncated Normal,
Student-t, Power Exponential, Pearson type VII, Slash, and Contaminated Normal
distributions, respectively.

The argument \code{nu} denotes a parameter or a vector of parameters depending
on the underlying density generating function (DGF). For the truncated
Student-t, Power Exponential, and Slash distributions, \code{nu} must be a
positive scalar. For the truncated Pearson type VII distribution, \code{nu}
is a vector of length two, where the first element must be greater than
\eqn{p/2} and the second element must be strictly positive. For the truncated
Contaminated Normal distribution, \code{nu} is a vector of length two with
components taking values in the interval \eqn{(0,1)}.

This function also supports random number generation from truncated elliptical
distributions not explicitly listed in the \code{dist} argument by supplying
the density generating function (DGF) through either the \code{expr} or
\code{gFun} arguments. The DGF must be a non-negative and strictly decreasing
function on \eqn{(0,\infty)}.

The simplest approach is to provide the DGF to the \code{expr} argument as a
character string. The notation used in \code{expr} must be compatible with both
the \pkg{Ryacas} package and the \pkg{R} evaluation environment. For example,
for the DGF \eqn{g(t) = e^{-t}}, the user should specify
\code{expr = "exp(-t)"}. The expression must depend only on the variable
\eqn{t}; any additional parameters must be supplied as fixed values.
When a character expression is provided via \code{expr}, the algorithm attempts
to compute a closed-form expression for the inverse of \eqn{g(t)}. Since such an
expression may not always exist, a warning is issued whenever the inversion
cannot be obtained analytically (see Example 2).

If random samples cannot be generated using a character expression supplied to
\code{expr}, the user may instead provide a custom \pkg{R} function through the
\code{gFun} argument. By default, the inverse of this function is approximated
numerically; however, for improved computational efficiency, the user may
optionally provide its inverse via the \code{ginvFun} argument. When \code{gFun}
is supplied, the arguments \code{dist} and \code{expr} are ignored.
}
\note{
The Normal distribution is a special case of the Power Exponential distribution
obtained when \code{nu = 1}. The Student-t distribution with \eqn{\nu} degrees of
freedom arises as a particular case of the Pearson type VII distribution when
\code{nu = } ((\eqn{\nu}+p)/2, \eqn{\nu}).
}
\examples{
library(ggplot2)
library(ggExtra)
library(gridExtra)

# Example 1: Sampling from the Truncated Normal distribution
set.seed(1234)
mu  = c(0, 1)
Sigma = matrix(c(1,0.70,0.70,3), 2, 2)
lower = c(-2, -3)
upper = c(3, 3)
sample1 = rtelliptical(5e4, mu, Sigma, lower, upper, dist="Normal")

# Histogram and density for variable 1
ggplot(data.frame(sample1), aes(x=X1)) +
   geom_histogram(aes(y=..density..), colour="black", fill="grey", bins=15) +
   geom_density(color="red") + labs(x=bquote(X[1]), y="Density") +
   theme_bw()

# Histogram and density for variable 2
ggplot(data.frame(sample1), aes(x=X2)) +
   geom_histogram(aes(y=..density..), colour="black", fill="grey", bins=15) +
   geom_density(color="red") + labs(x=bquote(X[2]), y="Density") +
   theme_bw()

\donttest{
# Example 2: Sampling from the Truncated Logistic distribution

# Function for plotting the sample autocorrelation using ggplot2
acf.plot = function(samples){
 p = ncol(samples); n = nrow(samples); q1 = qnorm(0.975)/sqrt(n); acf1 = list(p)
 for (i in 1:p){
   bacfdf = with(acf(samples[,i], plot=FALSE), data.frame(lag, acf))
   acf1[[i]] = ggplot(data=bacfdf, aes(x=lag,y=acf)) + geom_hline(aes(yintercept=0)) +
     geom_segment(aes(xend=lag, yend=0)) + labs(x="Lag", y="ACF", subtitle=bquote(X[.(i)])) +
     geom_hline(yintercept=c(q1,-q1), color="red", linetype="twodash") +
     theme_bw()
 }
 return (acf1)
}

set.seed(5678)
mu  = c(0, 0)
Sigma = matrix(c(1,0.70,0.70,1), 2, 2)
lower = c(-2, -2)
upper = c(3, 2)
# Sample autocorrelation with no thinning
sample2 = rtelliptical(2000, mu, Sigma, lower, upper, expr="exp(-t)/(1+exp(-t))^2")
grid.arrange(grobs=acf.plot(sample2), top="Logistic distribution with no thinning", nrow=1)

# Sample autocorrelation with thinning = 3
sample3 = rtelliptical(2000, mu, Sigma, lower, upper, expr="exp(-t)/(1+exp(-t))^2",
                       thinning=3)
grid.arrange(grobs=acf.plot(sample3), top="Logistic distribution with thinning = 3", nrow=1)


# Example 3: Sampling from the Truncated Kotz-type distribution
set.seed(5678)
mu  = c(0, 0)
Sigma = matrix(c(1,-0.5,-0.5,1), 2, 2)
lower = c(-2, -2)
upper = c(3, 2)
sample4 = rtelliptical(2000, mu, Sigma, lower, upper, gFun=function(t){t^(-1/2)*exp(-2*t^(1/4))})
f1 = ggplot(data.frame(sample4), aes(x=X1,y=X2)) + geom_point(size=0.50) +
     labs(x=expression(X[1]), y=expression(X[2]), subtitle="Kotz(2,1/4,1/2)") +
     theme_bw()
ggMarginal(f1, type="histogram", fill="grey")}

}
\references{
{
  \insertRef{fang2018symmetric}{relliptical}

  \insertRef{ho2012some}{relliptical}

  \insertRef{neal2003slice}{relliptical}

  \insertRef{robert2010introducing}{relliptical}

  \insertRef{valeriano2021moments}{relliptical}
}
}
\seealso{
\code{\link{mvtelliptical}}
}
\author{
Katherine L. Valeriano, Christian E. Galarza and Larissa A. Matos
}
